Separate async logic and the view
