+++ client/src/main/scripts/components/auth/LoginTypes.ts
@@ -1,0 +1,11 @@
+export interface LoginReq {
+  email: string;
+  kind: string;
+  redirect?: string;
+}
+
+export interface LoginRes {
+  title: string;
+  body: string;
+  btn: string;
+}
